refactor(properties): Split properties.rs into smaller modules

chore: fix the license header check
